+++ magik-language-server/src/main/java/nl/ramsolutions/sw/magik/languageserver/MagikSettings.java
@@ -124,7 +124,8 @@ public class MagikSettings {
         }
 
         final String overrideConfigFileStr = overrideConfigFile.getAsString();
-        if (overrideConfigFileStr == null) {
+        if (overrideConfigFileStr == null
+            || overrideConfigFileStr.isEmpty()) {
             return null;
         }